Finish na-preserving conversions and warn on remaining gaps
(cherry picked from commit 463125c31a59254f6268f91fd0f40f6cf6051964)

+++ pineforge_codegen/codegen/emit_top.py
@@ -1900,7 +1900,7 @@ class TopLevelEmitter:
                     PineType.FLOAT: "double",
                     PineType.BOOL: "bool",
                     PineType.STRING: "std::string",
-                    PineType.COLOR: "int",
+                    PineType.COLOR: "int64_t",
                 }[variant_pt]
             elif i < len(getattr(fi, "param_type_specs", [])) and fi.param_type_specs[i] is not None:
                 # Precise per-param TypeSpec (declared hint or call-site inference):
